Log reason for remove_host action failing
If removing a host from an aggregate fails, the API route handler
just returns a new error message without the original failure being
logged. This logs the error before returning the error response with
the generic message.

Alternative to this we could just put the exception message in the
response text but I figured people might bristle at changing the
(mostly unhelpful) error message so I went the non-controversial route.

Change-Id: I8b41f8313f74aa3766c13c1265f6ee1f0fc03f67