cond: rename 'recreate' var to 'evacuate'
Largely cribbed from Ib7276dd23f7c7ba7bd7afe37147a8479feb4b333.

When recreate=True, we're evacuating during the rebuild_instance method.

This is not obvious given the method name, so this change renames the
variable at the top of the method and uses the new 'evacuate' variable
throughout the rest.

We could rename the parameter with some RPC version dancing, but that
gets complicated.

Change-Id: I383c486ff1bd1a4bb30ec22779469af6e9308f7c