Create a controller for qga when SEV is used
When a guest agent is requested, it requires a virtio-serial controller.
If the controller is not created explicitly, it will be created by
libvirt. But if AMD SEV is also requested, the controller is expected
to use the iommu driver.

In order to achive that, the controller is created explicitly with the
required driver.

Change-Id: I47c248649b9d77e8bd7c5350fc84d89342be4623
Closes-Bug: 1845986