Remove super old br-<uuid> neutron network id compat code
This removes some really old compat code from when what was
originally neutron (quantum) would allow network ids with a br-
prefix for some reason. If I had to guess I'd guess it was some
rackspace quirk.

Anyway, with the neutron network data model [1] the id field is
restricted to 36 characters [2][3] so br-<uuid> isn't going to
happen in reality so we can drop this old compat code.

[1] https://opendev.org/openstack/neutron/src/tag/15.0.0/neutron/db/models_v2.py#L255
[2] https://opendev.org/openstack/neutron-lib/src/tag/1.29.1/neutron_lib/db/model_base.py#L68
[3] https://opendev.org/openstack/neutron-lib/src/tag/1.29.1/neutron_lib/db/constants.py#L20

Change-Id: I58411a40c7d2b0d7f50b8d9c29c748438b1e1420