Remove service_uuids_online_data_migration
A blocker migration was added in Train [1] to force
deployments to make sure they have completed the
services.uuid online migration (added in Pike). Now
that we're in Ussuri we can drop that online data
migration code.

Note that InstanceListWithServicesTestCase is removed
because the scenario is now invalid with the blocker
DB migration.

[1] I8927b8a4513dab242d34953d13dd2cc95393dc80

Change-Id: If77702f0c3212f904443f627037782f9ad7b3b55