Use named kwargs in compute.API.resize
The only kwarg passed to resize() is the auto_disk_config
kwarg during resize (not cold migrate). This expands it out
to a named kwarg to resolve the TODO. The same is done with
_check_auto_disk_config and as a result there is a hit to
rebuild as well, but the same TODO exists on rebuild() to use
named kwargs but that can be dealt with separately.

Change-Id: Ide8eb9e09d22f20165474d499ef0524aefc67854