Remove now invalid TODO from ComputeManager._confirm_resize
Back when I added the TODO [1] the libvirt driver was getting BDMs
from the instance to determine if it was a volume-backed server
and I figured when _confirm_resize eventually started sending
versioned notifications with BDMs we could kill two birds with one
DB query stone to get the BDMs.

However, [2] removed the need for the volume-backed instance check
in the driver which makes the TODO in the compute manager code
invalid now so this just removes the TODO comment.

[1] https://review.opendev.org/#/c/516395/2/nova/compute/manager.py@3630
[2] Ib10081150e125961cba19cfa821bddfac4614408

Change-Id: I107931f0817eab3dc030fef13d1c63426605f537