Implement update_provider_tree for mocked driver in test_resource_tracker
This resolves the TODO in the test_resource_tracker module by
implementing the update_provider_tree method on the mocked
virt driver. As a result a few tests need to mock out the call to
_sync_compute_service_disabled_trait since we no longer hit the
NotImplementedError block and skip that method.

The same is done in the functional test as well even though there
was no TODO for it in that module.

This is part of a series of changes to eventually drop compat for
non-update_provider_tree implementations.

Change-Id: Iff7805deb041596db30465b52658ca77ddf598dd