Remove TODO from ComputeTaskManager._live_migrate
Back when this TODO was written [1] the wrap_instance_event
decorator was not on the migrate_server method that calls
_live_migrate and the other caller, live_migrate_instance,
did not exist (but also has the wrap_instance_event decorator).
Because _live_migrate re-raises and both callers are using the
wrap_instance_event decorator which will update the instance
action created in the API live_migrate method, the TODO is
essentially already resolved.

[1] I4120156db1499dfd3ed22095e528787eb73d33a6

Change-Id: I4d5276a2f8dc185fdead4592e1801411566558cb