Functional reproduce for bug 1852207
This is a stable/pike only change as the bug 1852207 cannot be
reproduced on newer version than stable/pike due to feature [1] fixed
the issue.

This is a regression that was introduced by a bugfix [2] backport to
stable/pike.

This patch adds a functional test that reproduce the bug.

[1] https://review.opendev.org/#/c/511358/
[2] https://review.opendev.org/#/c/606161/

Related-Bug: #1852207

Change-Id: I518ea316bfebed2d0fdbcedea18c727b2834a550