Store instance action event exc_val fault details
The InstanceActionEvent.pack_action_event_finish method was
not storing the exc_val in a field that was actually part
of the data model so it was never stored ("message" isn't a
column in the instance_actions_events table, "details" is).

This formats and stores the exc_val, if provided, using the
same utility code that is used to record the message for
an instance fault record.

Eventually we can build on this in the os-instance-actions API
by exposing the fault details to the user without needing to
expose the traceback (like the server fault "details" traceback).

Co-Authored-By: Brin Zhang <zhangbailin@inspur.com>

Part of blueprint action-event-fault-details

Change-Id: Ie3e11b38aac251c3f8911ed57dc5e7503aa91301