block_device: Copy original volume_type when missing for snapshot based volumes
Attempts to launch an instance from an encrypted volume snapshot would
previously fail if a volume_type was not specified in the
block_device_mapping of the boot request.

To avoid such failures DriverVolSnapshotBlockDevice will now attempt to
lookup and use the volume_type of the original volume that the snapshot
is based on. This should allow the eventual volume creation request
based on the snapshot to succeed in cases where a volume_type is
required but not provided in the boot request.

Closes-Bug: #1853495
Change-Id: Ic749c49e227e41732dbe04acea303b303acd264a