block_device: Copy original volume_type when missing for snapshot based volumes
Attempts to launch an instance from an encrypted volume snapshot would
previously fail if a volume_type was not specified in the
block_device_mapping of the boot request.

To avoid such failures DriverVolSnapshotBlockDevice will now attempt to
lookup and use the volume_type of the original volume that the snapshot
is based on. This should allow the eventual volume creation request
based on the snapshot to succeed in cases where a volume_type is
required but not provided in the boot request.

Conflicts:
        nova/virt/block_device.py
        nova/tests/unit/virt/test_block_device.py

NOTE(lyarwood): Due to I4205c00311f389907dcc390869414687ac03b7f5 not
being present in stable/stein. As a result the call to the
volume_api.create doesn't proivide the snapshot argument as a keyword.

Closes-Bug: #1853495
Change-Id: Ic749c49e227e41732dbe04acea303b303acd264a
(cherry picked from commit 5679a0bf99c17ff336fa520c9cdfad5156c6c9d2)
(cherry picked from commit 059dc01ae027f00ffffbaa9ee8a3c958031829ce)