Fix false ERROR message at compute restart
If an empty compute is restarted a false ERROR message was printed in
the log as the placement report client does not distinguish between
error from placement from empty allocation dict from placement.

This patch changes get_allocations_for_resource_provider to return None
in case of error instead of an empty dict. This is in line with
@safe_connect that would make the call return None as well. The
_error_out_instances_whose_build_was_interrupted also is changed to check
for None instead of empty dict before reporting the ERROR. The only
other caller of get_allocations_for_resource_provider was already
checking for None and converting it to an empty dict so from that caller
perspective this is compatible change on the report client.

This is stable only change as get_allocations_for_resource_provider was
improved during stein[1] to raise on placement error.

[1]I020e7dc47efc79f8907b7bfb753ec779a8da69a1

Change-Id: I6042e493144d4d5a29ec6ab23ffed6b3e7f385fe
Closes-Bug: #1852759
(cherry picked from commit 64f797a0514b0276540d4f6c28cb290383088e35)