Pick NEXT_MIN libvirt/QEMU versions for "V" release
The last time we incremented versions for libvirt and QEMU was during
the "Stein" release[1].  For "Train" we didn't do any (although we
advertized).  The advertized NEXT_MIN_{LIBVIRT,QEMU} versions for
"Train" release were set to: libvirt 4.0.0 and QEMU 2.11.0; but we
actually didn't bump them for "Train"; we'll do that for "Ussuri".

But before we bump for "Ussuri", we need to pick NEXT_MIN versions for
the "V" release.  Based on the LibvirtDistroSupportMatrix[3], this patch
announces a version increment for libvirt and QEMU that will be applied
for the "V" release, with predicted minimum versions setting:

    libvirt to 5.0.0; and
       QEMU to 4.0.0

[1] https://opendev.org/openstack/nova/commit/0b1548a988 -- libvirt:
    Bump MIN_{LIBVIRT,QEMU}_VERSION for "Stein", 2019-01-22
[2] https://opendev.org/openstack/nova/commit/0b1548a988 -- Pick next
    minimum libvirt / QEMU versions for "T" release
[3] https://wiki.openstack.org/wiki/LibvirtDistroSupportMatrix

Change-Id: I3d3679705dcfdd4b0ef058b40769e4c3811a1472
Signed-off-by: Kashyap Chamarthy <kchamart@redhat.com>