libvirt: Bump MIN_{LIBVIRT,QEMU}_VERSION for "Ussuri"
We didn't bump the MIN_{LIBVIRT_QEMU} For the "Train" release, although
we advertized it[1].  Let's actually bump it for "Ussuri", while we're
early on in the development cycle to  give it enough "soaking time".

As part of this bump, fix the small fallout:

  - Remove the MIN_LIBVIRT_MDEV_SUPPORT version constant; fix tests.
  - Remove the "GOP" (Graphic Output Protocol) video model version
    constant from MIN_LIBVIRT_VIDEO_MODEL_VERSIONS; fix tests.
  - Fix the fucntional test:
    VGPUReshapeTests.test_create_servers_with_vgpu

The following version constants and related tests, which are now no
longer required, will be removed in separate patches:

    MIN_LIBVIRT_KVM_AARCH64_VERSION
    MIN_LIBVIRT_MULTIATTACH
    MIN_LIBVIRT_FILE_BACKED_VERSION
    MIN_QEMU_FILE_BACKED_VERSION
    MIN_QEMU_FILE_BACKED_DISCARD_VERSION

And here[2] is the discussion for picking future libvirt/QEMU versions
for "V" release on the list.

[1] https://opendev.org/openstack/nova/commit/489b5f762e
    -- Pick next minimum libvirt / QEMU versions for "T" release,
    2018-09-25)
[2] http://lists.openstack.org/pipermail/openstack-discuss/2019-November/010907.html
    -- On next minimum libvirt / QEMU versions for "V" release

Change-Id: Ia18e9be4d07b732afd9db0007c541253f3cdaf82
Signed-off-by: Kashyap Chamarthy <kchamart@redhat.com>