Flesh out RevertResizeTask.rollback
This implements the TODO in the RevertResizeTask.rollback
method by logging an exception with some details and also
setting the instance to ERROR state, recording a fault
and sending an error notification, similar to the rollback
method in ConfirmResizeTask.

As noted inline, the exception message could be more detailed
but I've left that for a later change in case people want more
details.

Part of blueprint cross-cell-resize

Change-Id: I41296696da3226767d2bacba9345d829529ce4b6