Add functional cross-cell revert test with detached volume
This implements one of the test todos where we create a server
and attach a volume to it, resize to another cell, detach the
volume while the instance is in VERIFY_RESIZE status, and then
revert the resize and make sure the volume is still detached
from the source cell after the revert.

A log message is added to prep_snapshot_based_resize_at_source
which was useful while debugging the new test.

Part of blueprint cross-cell-resize

Change-Id: I89aef20eebb817822d830db178527c370e532f54