Force config drive in nova-next multinode job
We made the nova-next job multinode in Ussuri and we
sometimes see intermittent failures setting up guests
in tempest tests that move a guest from one host to
another, e.g. test_shelve_volume_backed_instance. If
this is due to slow nodes or slow network connections
between the subnode and the primary where the metadata
API server is running, the guest setup can fail which
results in SSH failures in the tempest tests which are
generally not easy to debug.

This makes the nova-next job force the use of a config
drive for the servers that are created so we can take
the metadata API server out of the mix for these types
of SSH failures with this specific job.

Change-Id: Idc45d5ebc26e3a8a94fc5ec13c07465b9dd460e6
Related-Bug: #1853453