Add contributor doc for resize and cold migrate
Resize and cold migrate can be confusing since they are
very similar operations and share mostly the same code paths
but there are some notable differences. This adds a contributor
doc, similar to the evacuate-vs-rebuild contributor doc, to
try and explain things at a high level and provide enough links
so contributors can follow along and see where things fit into
the puzzle.

A sequence diagram is sorely needed for this as well but that will
come in a separate change.

Change-Id: I11b401a3f874226fdc20a0ee0bd518192f70fa1d