api-guide: remove empty sections about inter-service interactions
The todos here are just kind of embarrassing so rather than try
to flesh these out we should just remove them.

The alternative would be writing something at a very high level
like 'the compute service works with the image service to get
the guest operating system image and create snapshots', 'the
compute service works with the block storage service to use
persistent volumes as root volumes with an operating
system image or data volumes', 'the compute service works with
the networking service to provide network access to the server
and firewall rules using security groups' and then link to the
glance, cinder and neutron docs, but I'm not sure it's worth it.

Change-Id: I7f7ca4cc5781a93c0b7b508b5c29e1f6f28194ad