api-guide: flesh out flavor extra specs and image properties
This adds high level details and links to more detailed docs
in both nova and glance.

This also un-hides the document so it shows up in the table of
contents on the main page. Previously it was being linked from
the general info page and some typos in there are fixed as well.

Change-Id: I495b9ae7855d36cb2597272bcae7ec00360bbed4