Support live migration with qos ports
This patch enhances the live_migration task in the conductor to support
qos ports during live migration. The high level sequence of events are
the following:
* when request spec is gathered before the scheduler call the resource
requests are collected from neutron ports and the request spec is
updated
* after a successful scheduling the request group - resource provider
mapping is calculated
* the instance pci requests are updated to drive the pci claim on the
target host to allocate VFs from the same PCI PF the bandwidth is
allocated from
* the inactive port binding on the target host is updated to have the RP
UUID in the allocation key according to the resource allocation on the
destination host.

As the port binding is already updated in the conductor the late check
about the allocation key in the binding profile is turned off for live
migration in the neutronv2 api.

Note that this patch only handles the live migration without target host
subsequent patches will add support for migration with target host and
other edge case like reschedule.

blueprint: support-move-ops-with-qos-ports-ussuri

Change-Id: Ibb84ea210795634f02997d4627e0beec5fea36ee