Extend NeutronFixture to allow live migration with ports
None of our functional live migration tests uses neutron ports. Simply
adding a port to the boot request shows that the live migration step of
the tests fails because the compute manager's
_do_pre_live_migration_from_source() call wait for vif-plugged events
from Neutron but our Neutron fixture does not send it.

This does not cause problem during a simple boot because during boot the
virt driver triggers the waiting for the vif-plug event and in the
functional test we have a FakeDriver that does not do it.

This patch implements the simplest solution to make it possible to test
live migration with neutron ports. It simply turns off the waiting by
setting CONF.vif_plugging_timeout to 0

A nicer solution would be to extend either the Neutron fixture or the
fake virt driver to send vif-plug event at certain triggers. But that
would require a lot more surgery.

This patch is a prerequisite to later patches covering live migration
with qos neutron ports.

blueprint: support-move-ops-with-qos-ports-ussuri

Change-Id: I63c1109dcdb9132cdbc41010654c5fdb31a4fe31