nova-net: Remove SG tests that don't apply to neutron
The neutron security group driver tests subclass the nova-network ones.
Since there are some tests that don't make sense for neutron, these are
stubbed out for the tests for same. Simply remove them from both as a
another step towards unifying these test classes.

In addition, remove a test that does "work" with neutron but doesn't
make any sense, since it's testing that nova is validating something
when in reality neutron is doing the validation.

Change-Id: I967ba6abf38dd0e991a070c76decacc579bb4ff7
Signed-off-by: Stephen Finucane <sfinucan@redhat.com>