Pass the target in os-services APIs policy
Currently if target is not passed in context.can(),
it use defauls target which is context.user_id, context.project_id.
These defaults target are not useful as it pass the
context's user_id and project_id only which means we tell
oslo policy to verify the context data with context data.

This commit pass the actual target for os-services API policies
which is empty dict because policy rule is system scoped rather
than project, so the token scope check deals with the required
target checking.

blueprint policy-defaults-refresh

Change-Id: I575036854c0081e682bd2b7bceb5ed876d7a73c0