Process requested_resources in ResourceRequest init
Previously, resources_from_request_spec was initializing ResourceRequest
and then tacking on the RequestSpec.resource_requests. This made little
sense in light of the fact that ResourceRequest init's job is processing
the RequestSpec.

To make the flow more scrutable, this commit integrates the processing
of RequestSpec.requested_resources into ResourceRequest.__init__.

As a result, ResourceRequest.add_request_group() is no longer used
outside of ResourceRequest itself, so it is made private.

No test changes are necessary with this refactor. All the flows that
were previously calling ResourceRequest.__init__ directly were either
doing so with a RequestSpec with no resource_requests, or were ignoring
the resource_requests in ways that didn't matter (at least not yet). The
latter had two cases:

- isolate_aggregates. Theoretically if you had tried to isolate
  aggregates based on traits fed in through port bandwidth requests,
  those traits would have been ignored.
- Selecting hosts with specific CPU model flags (libvirt only).
  Theoretically your bandwidth request could have required certain CPU
  flag-based traits and they would have been ignored.

Change-Id: Ibb0ee191c76df0d34c83daf97f462380e2741c94